findbugs: fix problems
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@51026 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/io/MyXMLHandler.java
@@ -2197,9 +2197,9 @@ public class MyXMLHandler implements DocHandler {
 					? width : height);
 			double relative;
 			if (tmp_showSpreadsheet) {
-				relative = tmp_sp1 / total;
+				relative = tmp_sp1 / (double) total;
 			} else {
-				relative = tmp_sp2 / total;
+				relative = tmp_sp2 / (double) total;
 			}
 			spXml = new DockSplitPaneData[] {
 					new DockSplitPaneData("", relative, splitOrientation) };
